UBERF-7690: Use query joiner for server/trigger requests
