+++ server/middleware/src/queryJoin.ts
@@ -13,28 +13,32 @@
 // limitations under the License.
 //
 
-import { Class, Doc, DocumentQuery, FindOptions, FindResult, MeasureContext, Ref, Tx } from '@hcengineering/core'
-import { Middleware, SessionContext, TxMiddlewareResult, type ServerStorage } from '@hcengineering/server-core'
+import {
+  type Class,
+  type Doc,
+  DocumentQuery,
+  FindOptions,
+  FindResult,
+  type MeasureContext,
+  Ref,
+  type Tx
+} from '@hcengineering/core'
+import { Middleware, type ServerStorage, SessionContext, TxMiddlewareResult } from '@hcengineering/server-core'
 import { BaseMiddleware } from './base'
 
-import { deepEqual } from 'fast-equals'
+import { QueryJoiner } from '@hcengineering/server-core'
 
-interface Query {
-  _class: Ref<Class<Doc>>
-  query: DocumentQuery<Doc>
-  result: FindResult<Doc> | Promise<FindResult<Doc>> | undefined
-  options?: FindOptions<Doc>
-  callbacks: number
-  max: number
-}
 /**
  * @public
  */
 export class QueryJoinMiddleware extends BaseMiddleware implements Middleware {
-  private readonly queries: Map<Ref<Class<Doc>>, Query[]> = new Map<Ref<Class<Doc>>, Query[]>()
+  private readonly joiner: QueryJoiner
 
   private constructor (storage: ServerStorage, next?: Middleware) {
     super(storage, next)
+    this.joiner = new QueryJoiner((ctx, _class, query, options) => {
+      return storage.findAll(ctx, _class, query, options)
+    })
   }
 
   static async create (ctx: MeasureContext, storage: ServerStorage, next?: Middleware): Promise<QueryJoinMiddleware> {
@@ -52,59 +56,6 @@ export class QueryJoinMiddleware extends BaseMiddleware implements Middleware {
     options?: FindOptions<T>
   ): Promise<FindResult<T>> {
     // Will find a query or add + 1 to callbacks
-    const q = this.findQuery(_class, query, options) ?? this.createQuery(_class, query, options)
-    if (q.result === undefined) {
-      q.result = this.provideFindAll(ctx, _class, query, options)
-    }
-    if (q.result instanceof Promise) {
-      q.result = await q.result
-      q.callbacks--
-    }
-    this.removeFromQueue(q)
-
-    return q.result as FindResult<T>
-  }
-
-  private findQuery<T extends Doc>(
-    _class: Ref<Class<T>>,
-    query: DocumentQuery<T>,
-    options?: FindOptions<T>
-  ): Query | undefined {
-    const queries = this.queries.get(_class)
-    if (queries === undefined) return
-    for (const q of queries) {
-      if (!deepEqual(query, q.query) || !deepEqual(options, q.options)) {
-        continue
-      }
-      q.callbacks++
-      q.max++
-      return q
-    }
-  }
-
-  private createQuery<T extends Doc>(_class: Ref<Class<T>>, query: DocumentQuery<T>, options?: FindOptions<T>): Query {
-    const queries = this.queries.get(_class) ?? []
-    const q: Query = {
-      _class,
-      query,
-      result: undefined,
-      options: options as FindOptions<Doc>,
-      callbacks: 1,
-      max: 1
-    }
-
-    queries.push(q)
-    this.queries.set(_class, queries)
-    return q
-  }
-
-  private removeFromQueue (q: Query): void {
-    if (q.callbacks === 0) {
-      const queries = this.queries.get(q._class) ?? []
-      this.queries.set(
-        q._class,
-        queries.filter((it) => it !== q)
-      )
-    }
+    return await this.joiner.findAll(ctx.ctx, _class, query, options)
   }
 }
